+++ src/components/pages/apps/genes/gsea/ext-gsea/svg/genes-svg.tsx
@@ -1,10 +1,6 @@
 import { useCallback, useMemo, type ReactNode } from 'react'
 
-import {
-  axisDomainToRange,
-  axisDomainToRangeFunc,
-  axisLength,
-} from '@/components/plot/axes/axis'
+import { axisDomainToRangeFunc, axisLength } from '@/components/plot/axes/axis'
 import type { IGseaResult } from '@/lib/gsea/ext-gsea'
 
 import { useAxis } from '@/components/plot/axes/axes-store'
@@ -15,13 +11,11 @@ import { SvgText } from '@/components/plot/svg-text'
 import { IPos } from '@/interfaces/pos'
 import { COLOR_BLACK } from '@/lib/color/color'
 import { screenToSvgPoint, svgPointToScreen } from '@/lib/graphics/svg'
-import { geneSetScores, type IGeneSet } from '@/lib/gsea/geneset'
-import { abs } from '@/lib/math/abs'
+import { geneSetScores, IScoreGene, type IGeneSet } from '@/lib/gsea/geneset'
 import { max } from '@/lib/math/math'
 import { findNearest } from '@/lib/search'
 import { useCrosshair } from '@/providers/crosshair-provider'
 import { useSVG } from '@/providers/svg-provider'
-import { useTooltip } from '@/providers/tooltip-provider'
 import { IExtGseaPlotResult, useExtGseaContext } from '../ext-gsea-provider'
 import { IExtGseaSettings } from '../ext-gsea-settings'
 
@@ -37,7 +31,7 @@ export function ExtGseaHitsSvg({
   result: IExtGseaPlotResult
   gs: IGeneSet
   gsea: IGseaResult
-  scores: number[]
+  scores: IScoreGene[]
   maxScore: number
   gsMode: 'gs1' | 'gs2'
   pos: IPos
@@ -45,7 +39,7 @@ export function ExtGseaHitsSvg({
   const { plot } = useExtGseaContext()
   const { ref } = useSVG()
   const { showCrosshair, hideCrosshair } = useCrosshair()
-  const { showTooltip, hideTooltip } = useTooltip()
+  //const { showTooltip, hideTooltip } = useTooltip()
 
   const { axis: xax } = useAxis({
     plotId: result.id,
@@ -72,10 +66,10 @@ export function ExtGseaHitsSvg({
     return points
   }, [gsea.esHits, xax])
 
-  const _hideTooltip = useCallback(() => {
-    hideCrosshair()
-    hideTooltip()
-  }, [hideCrosshair, hideTooltip])
+  //   const _hideTooltip = useCallback(() => {
+  //     hideCrosshair()
+  //     hideTooltip()
+  //   }, [hideCrosshair, hideTooltip])
 
   const onMouseMove = useCallback(
     (e: React.MouseEvent) => {
@@ -99,8 +93,7 @@ export function ExtGseaHitsSvg({
         plotP.y < 0 ||
         plotP.y > displayProps.genes.height
       ) {
-        //setBarPos(null)
-        _hideTooltip()
+        hideCrosshair()
         return
       }
 
@@ -110,31 +103,41 @@ export function ExtGseaHitsSvg({
       )
 
       if (Math.abs(plotP.x - nearest) > 5) {
-        _hideTooltip()
+        hideCrosshair()
         return
       }
 
+      const rank = gsea.esHits[index].rank
+
       const barP = {
         x: nearest + displayProps.plot.margin.left + pos.x,
-        y: plotP.y + displayProps.plot.margin.top + pos.y,
+        y: displayProps.plot.margin.top + pos.y + displayProps.genes.height / 2,
       }
 
-      const { relativeP: barScreenP, screenP } = svgPointToScreen(
-        ref.current,
-        barP
-      )
+      const { relativeP: barScreenP } = svgPointToScreen(ref.current, barP)
 
-      showCrosshair(barScreenP)
-      showTooltip({
-        pos: { x: screenP.x + 5, y: screenP.y + 5 },
+      showCrosshair({
+        pos: barScreenP,
         content: (
           <>
-            <strong>{gs.name}</strong>
-            <span>{`Name: ${gsea.esHits[index].name}`}</span>
-            <span>{`Rank: ${gsea.esHits[index].rank.toLocaleString()}, Score: ${gsea.esHits[index].score.toFixed(3)}`}</span>
+            <strong>
+              {gsea.esHits[index].name} ({gs.name})
+            </strong>
+            <span>{`Score: ${result.rankedGenes[rank].score.toFixed(3)}`}</span>
+            <span>{`Rank: ${gsea.esHits[index].rank.toLocaleString()}`}</span>
           </>
         ),
       })
+      //   showTooltip({
+      //     pos: { x: screenP.x + 5, y: screenP.y + 5 },
+      //     content: (
+      //       <>
+      //         <strong>{gs.name}</strong>
+      //         <span>{`Name: ${gsea.esHits[index].name}`}</span>
+      //         <span>{`Rank: ${gsea.esHits[index].rank.toLocaleString()}, Score: ${gsea.esHits[index].score.toFixed(3)}`}</span>
+      //       </>
+      //     ),
+      //   })
     },
     [
       pos,
@@ -146,13 +149,15 @@ export function ExtGseaHitsSvg({
       result,
       showCrosshair,
       hideCrosshair,
-      showTooltip,
-      hideTooltip,
     ]
   )
 
   const genesSvg = useMemo(() => {
     let genesSvg: ReactNode | undefined = undefined
+
+    if (!points || points.length === 0) {
+      return null
+    }
 
     if (displayProps.genes.line.show) {
       // scale colors to score, for generic ext gsea
@@ -160,17 +165,19 @@ export function ExtGseaHitsSvg({
       // we can scale by strength of interaction with
       // target
 
-      let hitIdx = gsea.esHits.map((g) => g.rank)
+      //let hitIdx = gsea.esHits.map((g) => g.rank)
 
-      let xs = axisDomainToRange(xax, hitIdx)
+      //let xs = axisDomainToRange(xax, hitIdx)
+
+      //const xaf = axisDomainToRangeFunc(xax)
 
       return (
         <>
           <SvgG id="hits">
-            {hitIdx.map((hit, hiti) => {
-              const x = xs[hiti]
+            {scores.map((hit, hiti) => {
+              const x = points[hiti].x // ?? xaf(gsea.esHits[hiti].rank)
 
-              const score = scores[hiti] / maxScore
+              const score = hit.score / maxScore
               const diff = 1 - score
 
               //need to vary between 1 and score/max score according to the gene score
@@ -220,7 +227,7 @@ export function ExtGseaHitsSvg({
             fill="transparent"
             pointerEvents="all"
             onMouseMove={onMouseMove}
-            onMouseLeave={_hideTooltip}
+            onMouseLeave={hideCrosshair}
           />
         </>
       )
@@ -233,11 +240,12 @@ export function ExtGseaHitsSvg({
     gsea,
     xax,
     scores,
+    points,
     maxScore,
     displayProps,
     pos,
     onMouseMove,
-    _hideTooltip,
+    hideCrosshair,
   ])
 
   return genesSvg
@@ -264,13 +272,25 @@ export function ExtGseaGenesSvgPlot({
     let genesSvg: ReactNode | undefined = undefined
 
     if (displayProps.genes.line.show) {
-      const scores1 = abs(geneSetScores(gs1).map((g) => g.score))
-      const scores2 = abs(geneSetScores(gs2).map((g) => g.score))
+      // the score weights of each gene hit, not the es
+      const scores1 = geneSetScores(gs1).map((g) => ({
+        ...g,
+        score: Math.abs(g.score),
+      }))
+
+      const scores2 = geneSetScores(gs2).map((g) => ({
+        ...g,
+        score: Math.abs(g.score),
+      }))
+
       // scale colors to score, for generic ext gsea
       // score is always 1 so no effect, for viper
       // we can scale by strength of interaction with
       // target
-      let maxScore = max([...scores1, ...scores2])
+      let maxScore = max([
+        ...scores1.map((g) => g.score),
+        ...scores2.map((g) => g.score),
+      ])
 
       const yOffset = displayProps.genes.height + 0.25 * displayProps.plot.gap.y