+++ src/providers/crosshair-provider.tsx
@@ -1,11 +1,21 @@
+import { BaseCol } from '@/components/layout/base-col'
 import { IPos } from '@/interfaces/pos'
+import { cn } from '@/lib/shadcn-utils'
 import { ReactNode, useEffect } from 'react'
 import { create } from 'zustand'
 import { samePosition, TOOLTIP_CLEAR_MS } from './tooltip-provider'
 
+interface ICrosshair {
+  pos: IPos
+  content?: ReactNode
+  offset?: IPos
+}
+
+const DEFAULT_OFFSET: IPos = { x: 10, y: 10 }
+
 interface ICrosshairStore {
-  crosshair: IPos | null
-  showCrosshair: (pos: IPos | null) => void
+  crosshair: ICrosshair | null
+  showCrosshair: (crosshair: ICrosshair) => void
   hideCrosshair: () => void
   dispose: () => void
 }
@@ -13,7 +23,7 @@ interface ICrosshairStore {
 export const useCrosshairStore = create<ICrosshairStore>()((set, get) => {
   let clearTimeoutId: ReturnType<typeof setTimeout> | null = null
   let crosshairFrame: number | null = null
-  let pendingCrosshair: IPos | null = null
+  let pendingCrosshair: ICrosshair | null = null
 
   const cancelPendingFrame = () => {
     if (crosshairFrame !== null) {
@@ -27,7 +37,9 @@ export const useCrosshairStore = create<ICrosshairStore>()((set, get) => {
   return {
     crosshair: null,
 
-    showCrosshair: (pos) => {
+    showCrosshair: (crosshair) => {
+      const { pos, content, offset = DEFAULT_OFFSET } = crosshair
+
       if (clearTimeoutId) {
         clearTimeout(clearTimeoutId)
         clearTimeoutId = null
@@ -38,11 +50,11 @@ export const useCrosshairStore = create<ICrosshairStore>()((set, get) => {
       //   return
       // }
 
-      if (crosshairFrame !== null && samePosition(pendingCrosshair, pos)) {
+      if (crosshairFrame !== null && samePosition(pendingCrosshair.pos, pos)) {
         return
       }
 
-      pendingCrosshair = pos
+      pendingCrosshair = { pos, content, offset }
 
       if (crosshairFrame !== null) {
         return
@@ -54,7 +66,7 @@ export const useCrosshairStore = create<ICrosshairStore>()((set, get) => {
         const nextCrosshair = pendingCrosshair
         pendingCrosshair = null
 
-        if (!samePosition(get().crosshair, nextCrosshair)) {
+        if (!samePosition(get().crosshair?.pos, pos)) {
           set({ crosshair: nextCrosshair })
         }
       })
@@ -108,34 +120,62 @@ export function CrosshairProvider({ children }: { children?: ReactNode }) {
     <>
       {children && children}
 
-      {crosshair && (
+      {crosshair?.pos && (
         <>
           <span
             className="absolute z-(--z-modal) bg-foreground/50 pointer-events-none w-0.5 h-full top-0"
-            style={{ left: crosshair.x - 1, height: crosshair.y - 5 }}
+            style={{ left: crosshair.pos.x - 1, height: crosshair.pos.y - 5 }}
           />
 
           <span
             className="absolute z-(--z-modal) bg-foreground/50 pointer-events-none w-0.5 top-0 bottom-0"
-            style={{ left: crosshair.x - 1, top: crosshair.y + 5 }}
+            style={{ left: crosshair.pos.x - 1, top: crosshair.pos.y + 5 }}
           />
 
           <span
             className="absolute z-(--z-modal) bg-foreground/50 pointer-events-none h-0.5 w-full left-0"
-            style={{ top: crosshair.y - 1, width: crosshair.x - 5 }}
+            style={{ top: crosshair.pos.y - 1, width: crosshair.pos.x - 5 }}
           />
 
           <span
             className="absolute z-(--z-modal) bg-foreground/50 pointer-events-none h-0.5 right-0"
-            style={{ top: crosshair.y - 1, left: crosshair.x + 5 }}
+            style={{ top: crosshair.pos.y - 1, left: crosshair.pos.x + 5 }}
           />
 
           <span
             className="absolute z-(--z-modal) bg-foreground/50 pointer-events-none h-0.5 w-0.5 top-0 left-0"
-            style={{ left: crosshair.x - 1, top: crosshair.y - 1 }}
+            style={{ left: crosshair.pos.x - 1, top: crosshair.pos.y - 1 }}
           />
+
+          {crosshair?.content && (
+            <BaseCol
+              className={cn(
+                'absolute z-(--z-tooltip) rounded-lg bg-black/50 shadow-lg px-4 py-3 text-xs text-white pointer-events-none'
+              )}
+              style={{
+                left: crosshair.pos.x + crosshair.offset.x,
+                top: crosshair.pos.y + crosshair.offset.y,
+              }}
+            >
+              {crosshair.content}
+            </BaseCol>
+          )}
         </>
       )}
     </>
   )
+}
+
+{
+  /* <BaseCol
+                className={cn(
+                  'fixed z-(--z-tooltip) rounded-lg bg-black/50 shadow-lg px-4 py-3 text-xs text-white pointer-events-none'
+                )}
+                style={{
+                  left: crosshair.pos.x + crosshair.offset.x,
+                  top: crosshair.pos.y + crosshair.offset.y,
+                }}
+              >
+                {crosshair.content}
+              </BaseCol>, */
 }